requirements.txt
streamlit==1.42.2
Pillow==10.3.0
